Example run
Take a chain with 50 atoms and 1 kink by construction (in the middle,
vacancy).

Relax the config by a short Langevin damped simulations.

Then drag it and see the king moving. The parameters is set so that the
chain in depinned but doesn't slide too much in the given time. Should
run in under a minute.

The system is under-damped, so the atoms oscillate a bit. Also, the kink
reflects nicely at the chain edges, does a double slip and chainges its
topological charge.