Adjust temperature and remove useless params
Remove mass, epsilon and K, not needed in adimensional model based on g.

Temperature is weird. We use amplitude for random numbers
brand=sqrt(2kBTgamma/dt) but this gives a temperature of 0.75 when input is 1.
By rescaling brand=sqrt(4/3 * 2kBTgamma/dt) all works well... but it's
unclear why this is needed. Is the Scipy Runge-Kutta implemented in a
strange way?